cleaned up abit
